# openapi/stage1-v1.yaml
openapi: 3.1.0
info:
  title: K-COMOS Stage1 API
  version: 1.0.0
  description: Korean Commercial Development OS Stage 1 contract API
servers:
  - url: http://localhost:8000/v1
security:
  - bearerAuth: []
paths:
...
  /projects/{project_id}/manual-adjustments:
    post:
      tags: [manual_adjustments]
      summary: Create manual adjustment
      parameters:
        - $ref: '#/components/parameters/projectId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ManualAdjustmentCreateRequest'}
            example:
              module_name: pricing
              field_name: recommended_monthly_rent_krw
              target_entity_type: unit
              target_entity_id: 11111111-1111-1111-1111-111111111111
              after_value: 4800000
              reason_code: local_market_override
              reason_note: 최근 권리금 조건 반영
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ManualAdjustmentCreateResponse'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /projects/{project_id}/manual-adjustments/{adjustment_id}/reset:
    post:
      tags: [manual_adjustments]
      summary: Reset manual adjustment
      parameters:
        - $ref: '#/components/parameters/projectId'
        - in: path
          name: adjustment_id
          required: true
          schema: {type: string, format: uuid}
      responses:
        '202':
          description: Reset accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  reset: {type: boolean}
                  recompute_job_id: {type: string}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    projectId:
      in: path
      name: project_id
      required: true
      schema: {type: string, format: uuid}
  responses:
...
    Forbidden:
      description: Forbidden
...
  schemas:
...
    ManualAdjustmentCreateRequest:
      type: object
      required: [module_name, field_name, target_entity_type, target_entity_id, after_value, reason_code]
      properties:
        module_name: {type: string}
        field_name: {type: string}
        target_entity_type: {type: string}
        target_entity_id: {type: string, format: uuid}
        after_value: {}
        reason_code: {type: string}
        reason_note: {type: string}
    ManualAdjustmentCreateResponse:
      type: object
      properties:
        adjustment_id: {type: string, format: uuid}
        approval_required: {type: boolean}
        recompute_job_id: {type: string}
...